core: Add internal __cu__new() that allows passing a tag_tables instance
And make cu__new() have the same interface as before, but allocate the
tags table, hand it to __cu__new() and ask it to delete it at
cu__delete() time.

Next patches will introduce cu__new_with_tables() for using the tags
table from a base CU, so that we can add more types from an alternate
DWARF file, like with .dwz.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>