dwarf_loader: Pass 'struct conf_load' pointer to cu__hash() to decide if it should be added or not
When adding tags from alternate DWARF files handling
DW_TAG_imported_unit we will need to look at conf_load fields to see if
we should really hash in this specific dwarf_cu hashtable, as the key is
the DWARF offset and that is per DWARF file.

We process the alternate DWARF file (the .dwz) in its entirety, so any
non-hashed TAG we added when processing DW_TAG_imported_unit tags in the
main DWARF file will be there for us to find when looking up the
small_id.

We have one hashtable per DWARF file and when doing the small_id
adjustments for the tag's types and abstract_origins, we need to look at
the right hashtable to obtain the small id.

We use the same ptr_table for both the main DWARF file and the alternate
one, to join them.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>