dwarf_loader: Set tag->type (small_id) from dtag->type (DWARF offset) for enumerations
As DW_TAG_enumeration_type entries have DW_AT_type to cope with packed
enums, etc, see in this example of a split DWARF (alt DWARF, .dwz):

  $ readelf -wi /usr/lib/debug/lib64/libopenvswitch-2.17.so.0.0.0-2.17.0-3.fc36.x86_64.debug | grep -m1 -B1 -A4 dp_packet_source
   <1><64a>: Abbrev Number: 56 (DW_TAG_enumeration_type)
      <64b>   DW_AT_name        : (alt indirect string, offset: 0xe456) dp_packet_source
      <64f>   DW_AT_encoding    : 7	(unsigned)
      <650>   DW_AT_byte_size   : 1
      <651>   DW_AT_type        : <alt 0x5e>
   <1><655>: Abbrev Number: 45 (DW_TAG_base_type)
  $ readelf -wi /usr/lib/debug/.dwz/openvswitch-2.17.0-3.fc36.x86_64 | grep '<5e>' -A4
   <1><5e>: Abbrev Number: 57 (DW_TAG_base_type)
      <5f>   DW_AT_byte_size   : 1
      <60>   DW_AT_encoding    : 8	(unsigned char)
      <61>   DW_AT_name        : (indirect string, offset: 0x16b4a): unsigned char
   <1><65>: Abbrev Number: 0
  $

This will allow us to generate compilable source when all we have is a
enum without enumerators, as in the case above, see that the
DW_TAG_enumeration_type has no children of type DW_TAG_enumerator, it is
followed by a sibling, a DW_TAG_base_type. In this case all we can emit
as source code is 'enum dp_packet_source;' but that doesn't allow the
generated from debugging info to be compilable, see the next cset to
see how we instead will use the enumeration type and keep the
enumeration name just as a comment.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>